Phylogenetic bayesian generalized linear mixed models for community data

Conducting Bayesian generalized linear mixed models for community data using INLA.

Usage

CPR(formula,
    priors = NULL,
    df,
    VCV_sp,
    comm,
    family,
    optim.lambda = TRUE,
    original.VCV = TRUE,
    optim.control = NULL,
    control.compute = list(waic = TRUE, config = TRUE),
    prediction.df = "auto",
    predictedfittedresponse = "best_m",
    inla.rerun = 1, ...)

Arguments

formula

an INLA formula.

priors

prior specifications. If NULL default priors will be used. Use inla.list.models("prior") to see all prior available.

df

a data frame with the response and predictors

VCV_sp

a phylogenetic covariance matrix at species levels

comm

a species site matrix

family

error distribution in the glmm. Defaulted as "gaussian". Use inla.list.models("likelihood") to see all distribution available.

optim.lambda

TRUE/FALSE. If TRUE, the function will optimize the (\(\lambda\)).

original.VCV

TRUE/FALSE. If TRUE, the function will build a model based on the phylogenetic covariance matrix supplied by the user.

optim.control

additional arguments passing to optim.

control.compute

arguments for control.compute in the inla function.

prediction.df

model predictions. If "auto", the model will provide predictions of a focal relatonship (e.g. ecosystem function and biodiversity) by holding other covariates at their mean (continuous) or mode (nominal) values. Users can also provide a data frame and ask INLA to make model predictions. The data frame must have the same structure as the analyzed datset, and users can define the conditions of different X while setting the response as NA.

predictedfittedresponse

compute the median posterior predictions of the best model ("best_m"), models with the phylogenetic covariance matrix supplied by users ("original_VCV"), or models without phylogenetic effect ("no_phylo"). The results can then passed to createDHARMa. Defaulted as "best_m".

inla.rerun

number of times to rerun INLA models to ensure stable results. Defaulted as 1.

...

additional arguments passing to inla

Details

When specifying the random effect, use Cmatrix = Phylo to indicate the phylogenetic covariance between communities, and comm as the identity of each community. See examples. You don't need to create them in the data frame. CPR will automatically create these for you in the internal environment.

The model will always compare the wAIC between the model without considering phylogeny and the model with optimized (\(\lambda\)). The model without phylogeny will be selected as "best model" if the wAIC does not reduce more than 2 after including the optimized/original \(\lambda\), which means $$wAIC_{optimized/original} - wAIC_{without\_phylogeny} > -2$$

Note that even if optimized (\(\lambda\)) = 0, the model is not equivalent to a linear regression, as the covariance matrices will still consider species overlap between communities. Therefore, the off-diagnoals in glmm are still not zeroes.

Value

best_model

INLA output for the best model

best_model_name

The model identified as the "best".

predictedfittedresponse

median posterior predictions

wAIC

wAIC of models (without phylogeny, with the user supplied phylogenetic covariance matrix, and with the optimized phylogeny).

R2

R-squared calculated as correlations between the observed responses and the mean posterior predictions

optim_lambda

The value of the optimized \(\lambda\).

initial_formula

initial formula provided by users.

without_phylo_model

results on the fixed effects from the model without phylogeny.

optimized_model

results on the fixed effects from the model after optimizing phylogeny.

original_VCV_model

results on the fixed effects from the model based on the covariance matrix provided by users.

Cmatrix

The phylogentic matrix in the optimization model. Note that this is a precision matrix instead of a covariance matrix to fulfill the requirement of INLA

prediction

predictted values for new data.

Author

Toby P.N. Tsang

See also

Examples

library(phytools)
library(tidyverse)
library(CPR)
data(KSR)
data(KSR_MLtree)
data(KSR_EF)

VCV_sp <- vcv(KSR_MLtree)

inla_m <- CPR(formula = bugs~Real.rich+f(comm,model="generic0",Cmatrix=Phylo),
              priors=NULL,
              df = KSR_EF,
              VCV_sp = VCV_sp,
              comm=KSR,
              family="nbinomial",
              optim.lambda = T)

summary(inla_m$best_model)
inla_m$prediction #see predicted relationships at each unique value of the a covariate, while holding others at the mean value.
inla_m$wAIC #wAIC of different models. Note that wAIC_optim has the lowest wAIC.