Generalized least squares regression for community data.

This function conducts GLS for community-level data. This is not recommended because pglmm can better control type-I error, power, and slope estimates. However, for some analyses (e.g., piecewiseSEMs), INLA is not supported yet, so a compromise is to use GLS.

Usage

CPR_GLS(formula,
        df,
        VCV_sp,
        comm,
        optim.lambda = TRUE,
        original.VCV = TRUE,
        ...)

Arguments

formula

a gls formula.

df

a data frame.

VCV_sp

a covariance matrix at species levels.

comm

a species-site matrix.

optim.lambda

TRUE/FALSE. If TRUE the optimial lambda and model will also be returned.

original.VCV

TRUE/FALSE. If TRUE the model using the supplied VCV_sp will also be returned.

...

arguments passing to optim.

Details

The function calls gls from nlme to conduct the gls, but will automatically calculate the covariance matrix between communities using get_comm_pair_r.

Value

without_phylo_model

The model without considering phylogeny (i.e., equivalent to an OLS regression).

optimized_model

The model with optimized \(\lambda\).

original_VCV_model

The model based on the phylogenetic covariance matrix provided by users (i.e., without optimization).

AIC

AIC of each model.

optim_lambda

the optimized \(\lambda\) (i.e., the \(\lambda\) leading to lowest AIC).

Author

Toby P.N. Tsang

See also

Examples

library(phytools)
library(tidyverse)
library(CPR)
data(KSR)
data(KSR_MLtree)
data(KSR_EF)

VCV_sp <- vcv(KSR_MLtree)

GLS_lambda_1 <- CPR_GLS(formula = LAI~Real.rich,
                        df = KSR_EF,
                        VCV_sp = VCV_sp,
                        optim.lambda = TRUE,
                        comm = KSR)

GLS_lambda_1$without_phylo_model #models without considering phylogeny
GLS_lambda_1$optimized_model #models without considering phylogeny
GLS_lambda_1$original_VCV_model #models without considering phylogeny
GLS_lambda_1$AIC #AIC of different models
GLS_lambda_1$optim_lambda #optimized lambda if available