Simulating responses from INLA models

Model diagnostics using DHARMa require simulated responses (see createDHARMa). This function automatically simulate the responses from an INLA model. Currently only support models of gaussian, poisson, binomial, negative binomial (type 0), tweedie, and zero-inflated negative binomial (type 1) model.

Usage

INLA_simulate(object,
              nsim = 250,
              Ntrails = NULL)

Arguments

object

INLA or CPR output (e.g. m$best_model)

nsim

Number of simulations

Ntrails

The number of trials for the binomial likelihood. This should match the input to the argument Ntrials in inla. Defaulted as 1.

Details

The simulated response can be passed to createDHARMa, creating a DHARMa object for further diagnostic.

Value

A data frame containing simulated responses, with each column representing each simulation run.

Author

Toby P.N. Tsang

See also

Examples

library(phytools)
library(tidyverse)
library(CPR)
data("KSR")
data("KSR_MLtree")
data("KSR_EF")

VCV_sp <- vcv(KSR_MLtree) #species level phyologenetic covariance matrix using default (Brownian) model
VCV_sp <- VCV_sp[order(rownames(VCV_sp)),order(colnames(VCV_sp))]

inla_m <- CPR(formula = bugs~Real.rich+f(comm,model="generic0",Cmatrix=Phylo),
              priors=NULL,
              df = KSR_EF,
              VCV_sp = VCV_sp,
              comm=KSR,
              family="nbinomial",
              optim.lambda = T)

library(DHARMa)
sres <- INLA_simulate(inla_m$best_model)
DHARMa_res <- createDHARMa(sres,
                           KSR_EF$bugs,
                           fittedPredictedResponse=inla_m$predictedfittedresponse,
                           integerResponse = T)
plot(DHARMa_res,quantreg=F)