Control Net Reduction

Run the Control Net Reduction Algorithm.

Usage

cnr(x, margin, n_polycoef = 20L, progress = interactive(), ...)

# S3 method for cpr_cnr
summary(object, ...)

Arguments

x

a cnr_cn object

margin

the margins to apply the CNR algorithm to. Passed to influence_weights.

n_polycoef

the number of polynomial coefficients to use when assessing the influence of each internal knot.

progress

show a progress bar.

...

not currently used

object

a cpr_cnr object

Details

cnr runs the control net reduction algorithm.

keep will keep the regression fit as part of the cnr\_cp object for models with up to and including keep fits. For example, if keep = 10 then the resulting cnr\_cnr object will have the regression fit stored in the first keep + 1 (zero internal knots, one internal knot, ..., keep internal knots) cnr\_cp objects in the list. The limit on the number of stored regression fits is to keep memory usage down.

See also