cpr: Control Polygon Reduction

The cpr package implements the control polygon reduction and control net reduction methods for finding parsimonious B-spline regression models as described in DeWitt (2017).

References

DeWitt, Peter Edward. “Parsimonious B-spline regression models via control polygon and control net reduction for identifying factors explaining variation in daily hormone profiles during the menopausal transition.” (2017).

See also

Author

Maintainer: Peter DeWitt dewittpe@gmail.com (ORCID)

Other contributors:

Samantha MaWhinney sam.mawhinney@ucdenver.edu [thesis advisor]

Nichole Carlson nichole.carlson@ucdenver.edu [thesis advisor]