Provide marginal and conditional R-squared for outputs from spaMM

Provide marginal and conditional R-squared for outputs from spaMM

Usage

get_R2(spaMM_m,spaMM_m_int_only = NULL)

Arguments

spaMM_m

model output from spaMM

spaMM_m_int_only

intercept-only model from spaMM. Only needed for negative binomial distribution.

Details

The function is based on R2 from the performance package, which does not support spaMM. Note that the model currently only supports gassuian and negative binomial (negbin2) distribution. If the model doesn't contain any compositional autocorrelation, it's probably easier to use other packages or functions (e.g. lme4, glmmTMB, lm, glm), which is better supported by other R packages.

You can also calculate the correlation coefficient between observed and predicted responses (prediction can be based on fixed effects only or full model, depending on the purposes). This is technically not the same as marginal and conditional R-squared, but should be a easy solution if you have models based on distributions not supported here.

Value

R2m

Marginal R-squared.

R2c

Conditional R-squared.

Author

Toby P.N. Tsang

See also

Examples