Getting pairwise phylogenetic correlation between communities

Getting pairwise phylogenetic correlation between communities

Usage

get_comm_pair_r(comm,V,force.PD = TRUE)

Arguments

comm

a species-site matrix

V

a phylogenetic covariance matrix at species levels

force.PD

find the closest positive definite matrix at community levels. Default as TRUE. Note that if matrix is not positive definite subsequent analyses can fail.

Details

The resulting phylogenetic correlation matrix is based on the additive law of covariance. Imagine two communities with species (X,Y) and species (X,Z). Covariance at community levels is $$Cov(X+Y,X+Z) = Var(X)+Cov(X,Z)+Cov(X,Y)+Cov(Y,Z)$$ All of the components can be obtained from the species level phylogenetic matrix.

Value

a phylogenetic correlation matrix at community levels.

Author

Toby P.N. Tsang

Examples

library(phytools)
library(tidyverse)
library(CPR)
data(KSR)
data(KSR_MLtree)
data(KSR_EF)

KSR_MLtree <- keep.tip(KSR_MLtree,colnames(KSR))
VCV_sp <- vcv(KSR_MLtree) #species level phyologenetic covariance matrix using default (Brownian) model
VCV_comm <- get_comm_pair_r(KSR,VCV_sp)