Getting pairwise phylogenetic correlation between communities

Getting pairwise phylogenetic correlation between communities

Usage

get_comm_pair_r(comm,V,force.PD = TRUE)

Arguments

comm

a species-site matrix

V

a phylogenetic covariance matrix at species levels

force.PD

find the closest positive definite matrix at community levels. Default as TRUE. Note that if matrix is not positive definite subsequent analyses can fail.

Details

The resulting phylogenetic correlation matrix is based on the additive law of covariance. Imagine two communities with species (X,Y) and species (X,Z). Covariance at community levels is Cov(X+Y,X+Z) = Var(X)+Cov(X,Z)+Cov(X,Y)+Cov(Y,Z). All of the components can be obtained from the species level phylogenetic matrix.

Value

a phylogenetic correlation matrix at community levels.

References

Author

Toby P.N. Tsang

Examples

library(phytools)
#> Loading required package: ape
#> Loading required package: maps
library(tidyverse)
#> ── Attaching core tidyverse packages ──────────────────────── tidyverse 2.0.0 ──
#> ✔ dplyr     1.1.4     ✔ readr     2.1.4
#> ✔ forcats   1.0.0     ✔ stringr   1.5.1
#> ✔ ggplot2   3.4.4     ✔ tibble    3.2.1
#> ✔ lubridate 1.9.2     ✔ tidyr     1.3.0
#> ✔ purrr     1.0.2     
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()
#> ✖ purrr::map()    masks maps::map()
#> ✖ dplyr::where()  masks ape::where()
#> ℹ Use the conflicted package (<http://conflicted.r-lib.org/>) to force all conflicts to become errors
library(CPR)
data(KSR)
data(KSR_MLtree)
data(KSR_EF)

KSF_EF <- KSR_EF 
excluded <- anti_join(KSR, KSR_EF, by = c("PLOT" = "Plot"))
KSR <- semi_join(KSR, KSR_EF, by = c("PLOT" = "Plot"))

KSR_misc <- KSR$PLOT
KSR <- KSR[,-1]
KSR[is.na(KSR)] <- 0
KSR <- KSR[rowSums(KSR) > 0,]

KSR_MLtree <- keep.tip(KSR_MLtree,colnames(KSR))
VCV_sp <- vcv(KSR_MLtree) #species level phyologenetic covariance matrix using default (Brownian) model
VCV_comm <- get_comm_pair_r(KSR,VCV_sp)
#> Loading required package: Matrix
#> 
#> Attaching package: 'Matrix'
#> The following objects are masked from 'package:tidyr':
#> 
#>     expand, pack, unpack
#> Loading required package: matrixcalc